backend/requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
boto3==1.21.16
botocore==1.24.16
cachetools==4.2.2
certifi==2021.10.8
colorama==0.4.4
Django==4.0.2
django-cors-headers==3.11.0
django-extensions==3.1.5
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
iniconfig==1.1.1
jmespath==0.10.0
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
whitenoise==5.3.0